[Java] Add AbstractMutableDirectBuffer and refactor existing buffers to extend it.
